chore: rename sidebars to sidebar
I missed these renamings in the previous commit when removing multiple
sidebar logic.
