+++ packages/common-all/src/index.ts
@@ -33,7 +33,7 @@ export * from "./store";
 export * from "./engine";
 export * from "./drivers";
 export * from "./DLogger";
-export * from "./sidebars";
+export * from "./sidebar";
 
 export { axios, AxiosError };
 export { DateTime };
